+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Extended the `Stack` models allowing for non-symmetric coupling between devices.
 - `Stack` current drivers can now be of any type are adequately scaled.
+- Custom definition of the `ScalarDriver` is now possible and documented.
 
 # 1.5.0-1.5.4